requirements.txt
django
djangorestframework
requests
django-environ
bcrypt
celery
djangorestframework-simplejwt